+++ analyzer/language/dotnet/nuget/nuget_test.go
