+++ analyzer/language/dotnet/nuget/testdata/invalid.txt
