+++ analyzer/language/dotnet/nuget/testdata/packages.config
