+++ analyzer/language/golang/binary/testdata/executable_gobinary
