+++ analyzer/language/golang/mod/testdata/gomod_many.sum
