--- analyzer/library/bundler/bundler.go
@@ -1,47 +1,0 @@
-package bundler
-
-import (
-	"os"
-	"path/filepath"
-
-	"golang.org/x/xerrors"
-
-	"github.com/aquasecurity/fanal/analyzer"
-	"github.com/aquasecurity/fanal/analyzer/library"
-	"github.com/aquasecurity/fanal/types"
-	"github.com/aquasecurity/fanal/utils"
-	"github.com/aquasecurity/go-dep-parser/pkg/bundler"
-)
-
-func init() {
-	analyzer.RegisterAnalyzer(&bundlerLibraryAnalyzer{})
-}
-
-const version = 1
-
-var (
-	requiredFiles = []string{"Gemfile.lock"}
-)
-
-type bundlerLibraryAnalyzer struct{}
-
-func (a bundlerLibraryAnalyzer) Analyze(target analyzer.AnalysisTarget) (*analyzer.AnalysisResult, error) {
-	res, err := library.Analyze(types.Bundler, target.FilePath, target.Content, bundler.Parse)
-	if err != nil {
-		return nil, xerrors.Errorf("unable to parse Gemfile.lock: %w", err)
-	}
-	return res, nil
-}
-
-func (a bundlerLibraryAnalyzer) Required(filePath string, _ os.FileInfo) bool {
-	fileName := filepath.Base(filePath)
-	return utils.StringInSlice(fileName, requiredFiles)
-}
-
-func (a bundlerLibraryAnalyzer) Type() analyzer.Type {
-	return analyzer.TypeBundler
-}
-
-func (a bundlerLibraryAnalyzer) Version() int {
-	return version
-}
